+++ src/FluxFlow.Engine/Scenarios/ScenarioRunStatus.cs
@@ -1,8 +1,0 @@
-namespace FluxFlow.Engine.Scenarios;
-
-public enum ScenarioRunStatus
-{
-    Passed,
-    Failed,
-    Canceled
-}