+++ src/FluxFlow.Engine/Scenarios/ScenarioStepRunStatus.cs
@@ -1,9 +1,0 @@
-namespace FluxFlow.Engine.Scenarios;
-
-public enum ScenarioStepRunStatus
-{
-    Passed,
-    Failed,
-    TimedOut,
-    Canceled
-}